feat: add limit -1 for query (#3658)
The query `size` works like this, it should more clear now:
1. size > 0, it will return hits equal to `size`
2. size = 0, it will return empty hits, this used to other purposes,
like: get total, get aggs.
3. size = -1, it will use a backend default limit to return hits,
setting by `ZO_QUERY_DEFAULT_LIMIT=1000`

+++ src/handler/http/request/traces/mod.rs
@@ -200,10 +200,10 @@ pub async fn get_latest_traces(
     };
     let from = query
         .get("from")
-        .map_or(0, |v| v.parse::<usize>().unwrap_or(0));
+        .map_or(0, |v| v.parse::<i64>().unwrap_or(0));
     let size = query
         .get("size")
-        .map_or(10, |v| v.parse::<usize>().unwrap_or(10));
+        .map_or(10, |v| v.parse::<i64>().unwrap_or(10));
     let mut start_time = query
         .get("start_time")
         .map_or(0, |v| v.parse::<i64>().unwrap_or(0));
@@ -419,7 +419,7 @@ pub async fn get_latest_traces(
             }
         };
 
-        let resp_size = resp_search.hits.len();
+        let resp_size = resp_search.hits.len() as i64;
         for item in resp_search.hits {
             let trace_id = item.get("trace_id").unwrap().as_str().unwrap().to_string();
             let trace_start_time = item.get("start_time").unwrap().as_i64().unwrap();
